表示改善 ファイルのコピー先を変更 開始時のファイル削除を選択可能にした

+++ ex_function.sh
@@ -7,67 +7,89 @@ trap 'ex_final' 3
 DEST=/opt/snk
 
 #実験準備
-ex_erase () {
-	echo "●データ消去"
-	echo " "
+ex_prepare () {
+        echo
+	echo "●実験準備"
 
-	#データの消去
-	supervisorctl stop all
-	supervisorctl remove optimizer
-	rm -f $DEST/output/*.log
-	rm -f $DEST/output/*.json
-	rm -f $DEST/data/*.json
-	echo "●/data・・outputの中のlog・jsonを消去しました"
-	echo " "
-	supervisorctl start cyclic controller
+	#optimizerの停止
+	if [ "$(supervisorctl status all|grep ^optimizer)" != "" ]; then
+	    supervisorctl stop optimizer
+	    supervisorctl remove optimizer
+	    echo "optimizerを停止しました"
+	fi
+	#
 
-	#システム再起動用
-	sleep 30s
+	#データ消去(指定があれば)
+	if [ "$1" != "" ]; then
+	    #デーモン停止
+	    supervisorctl stop cyclic controller
+
+	    #データの消去
+	    rm -f $DEST/output/*.log
+	    rm -f $DEST/output/*.json
+	    rm -f $DEST/data/*.json
+	    echo "outputの中のlog・jsonを消去しました"
+
+	    #デーモン再起動
+	    supervisorctl start cyclic controller
+	    echo "デーモン起動待ち"
+	    while [ "$(supervisorctl status cyclic|grep RUNNING)" == "" ]; do
+		sleep 1s
+	    done
+	    while [ "$(supervisorctl status controller|grep RUNNING)" == "" ]; do
+		sleep 1s
+	    done
+	    echo "デーモンが起動しました"
+	fi
 }
 
 #暖機運転開始
 ex_warm () {
+	echo
 	echo "●暖気運転を開始"
-	echo " "
 
 	#PID1,PID2をMNモードに変更
 	mosquitto_pub -h localhost -t snk/1 -m '{"PID1-AT/MT-remote": 1}'
 	mosquitto_pub -h localhost -t snk/1 -m '{"PID2-AT/MT-remote": 1}'
 	mosquitto_pub -h localhost -t snk/1 -m '{"PID1-value-remote": 1.0}'
 	mosquitto_pub -h localhost -t snk/1 -m '{"PID2-value-remote": 0.0}'
-	echo "●ヒーター出口制御100%、入口制御0%に設定しました"
-	echo " "
+	echo "ヒーター出口制御100%、入口制御0%に設定しました"
 
 	#電源、ヒーターON
 	mosquitto_pub -h localhost -t snk/1 -m '{"on/off-remote": 1.0}'
 	mosquitto_pub -h localhost -t snk/1 -m '{"Heater-value-remote": 100}'
-	echo "●電源、ヒーターをonにしました"
-	echo " "
+	echo "電源、ヒーターをonにしました"
 }
 
 # 指定した時間だけ待つ
 ex_wait () {
+    echo
+    echo "●待ち時間"
     echo "$1待ちます"
     sleep "$1"
 }
 
 #PID設定温度(目標温度SV)の設定
 ex_pidSV () {
+        echo
+	echo "●PID設定温度(目標温度SV)を設定"
 	mosquitto_pub -h localhost -t snk/1 -m '{"r": $1}'
-	echo "PID設定温度(目標温度SV)を設定"
 	echo "r:$1℃"
 }
 
 #PID設定温度(目標温度SV)の設定(2系統ある場合)
 ex_pidSV2 () {
+	echo
+	echo "●PID設定温度(目標温度SV)を設定"
 	mosquitto_pub -h localhost -t snk/1 -m '{"r": $1}'
         mosquitto_pub -h localhost -t snk/1 -m '{"r1": $2}'
-	echo "PID設定温度(目標温度SV)を設定"
 	echo "r:$1℃ $r1:$2℃"
 }
 
 #PID制御値をランダムに選別
 ex_pidrandam () {
+        echo
+        echo "●PID制御設定値を変更します"
 	#Kpの制御設定
 	randKp=`shuf -i 1-50000 -n1`
 	fewKp1=`shuf -i 1-99 -n1`
@@ -85,30 +107,28 @@ ex_pidrandam () {
 	mosquitto_pub -h localhost -t snk/1 -m "{\"Kp\": $randKp1}"
 	mosquitto_pub -h localhost -t snk/1 -m "{\"Td\": $randTd1}"
 	mosquitto_pub -h localhost -t snk/1 -m "{\"Ti\": $randTi1}"
-        echo "●PID制御設定値が変更されました"
         echo "Kp:$randKp1, Td:$randTd1, Ti:$randTi1"
-        echo " "
 }
 
 #PID制御設定値の変更
 ex_pid () {
+	echo
+        echo "●PID制御設定値を変更します"
 	#Kpの制御設定値
 	mosquitto_pub -h localhost -t snk/1 -m "{\"Kp\": $1}"
 	#Tdの制御設定値
 	mosquitto_pub -h localhost -t snk/1 -m "{\"Td\": $2}"
 	#Tiの制御設定値
 	mosquitto_pub -h localhost -t snk/1 -m "{\"Ti\": $3}"
-	echo "●PID制御設定値が変更されました"
 	echo "Kp:$1, Td:$2, Ti:$3"
-	echo " "
 }
 
 #ヒーター出口温度の上昇
 ex_tempup () {
 	temp=$1
 
+	echo
 	echo "●PID1の計測温度(y)を$temp℃まで上げます"
-	echo " "
 
 	y=`mosquitto_sub -h localhost -t snk/0 -C 1|sed -e 's/,/\n/g'|grep '"y"'|awk '{print $2}'`
 	until [ `echo "$y >= $temp"|bc` == 1 ]
@@ -116,19 +136,19 @@ ex_tempup () {
     		y=`mosquitto_sub -h localhost -t snk/0 -C 1|sed -e 's/,/\n/g'|grep '"y"'|awk '{print $2}'`
 	done
 
-	echo "●PID1の計測温度(y)が$temp℃まで上昇しました"
-	echo " "
+	echo "PID1の計測温度(y)が$temp℃まで上昇しました"
 }
 
 #モード切替、実験開始時刻記録
 ex_time () {
 	#atモードに切り替え
+	echo
 	echo "●ATモードに切り替え"
-	echo " "
         mosquitto_pub -h localhost -t snk/1 -m '{"PID1-AT/MT-remote": 0}'
         mosquitto_pub -h localhost -t snk/1 -m '{"PID2-AT/MT-remote": 0}'
 
 	#実験開始時刻
+	echo
 	echo "●実験開始"
 	time2=`mosquitto_sub -h localhost -t snk/0 -C 1|sed -e 's/,/\n/g'|grep '"time"'|awk '{print $2}'`
 	time3=${time2/%?/}
@@ -172,6 +192,7 @@ ex_start () {
 
 #実験終了
 ex_fin () {
+        echo
 	echo "●実験終了"
 	time2=`mosquitto_sub -h localhost -t snk/0 -C 1|sed -e 's/,/\n/g'|grep '"time"'|awk '{print $2}'`
 	time3=${time2/%?/}
@@ -179,7 +200,7 @@ ex_fin () {
 
 	#ファイル保存
 	time=`date '+%Y%m%d_%H%M%S'`
-	cp $DEST/output/data.json $DEST/output/data_${time}.json
+	cp $DEST/output/data.json data_${time}.json
 	echo "実験終了時刻:$time4"
 	mosquitto_pub -h localhost -t snk/1 -m '{"Heater-value-remote": 0}'
 	mosquitto_pub -h localhost -t snk/1 -m '{"on/off-remote": 0}'
